do same for tests
